requirements.txt
torch
torchvision
pillow
numpy
albumentations
opencv-python-headless